Some fixes, refactoring. Make case 5 logic

+++ src/main/java/org/controller/WriterController.java
@@ -8,33 +8,37 @@ import java.util.List;
 public class WriterController {
     public final WriterRepository writerRepository;
 
-    public WriterController (WriterRepository writerRepository){
+    public WriterController(WriterRepository writerRepository) {
         this.writerRepository = writerRepository;
     }
 
-    public Writer getWriterById (Integer id){
+    public Writer getWriterById(Integer id) {
         return writerRepository.getById(id);
     }
 
-    public List<Writer> getAllWriters(){
+    public List<Writer> getAllWriters() {
         return writerRepository.getAll();
     }
 
-    public Writer createWriter(Writer writer){
-        return writerRepository.save(writer);
+    public void createWriter(Writer writer) {
+        writerRepository.save(writer);
     }
 
-    public Writer updateWriter(Writer writer){
-        return writerRepository.update(writer);
+    public void updateWriter(Writer writer) {
+        writerRepository.update(writer);
     }
 
-    public void deleteWriterById(Integer id){
-        writerRepository.deleteById(id);
+    public boolean deleteWriterById(Integer id) {
+        if (writerRepository.existsById(id)) {
+            writerRepository.deleteById(id);
+            return true;
+        } else {
+            return false;
+        }
     }
 
-            public boolean isUniqueID(int id){
-                List<Writer> writers = writerRepository.getAll();
-                return writers.stream().noneMatch(writer -> writer.getId() == id);
-            }
-
+    public boolean isUniqueID(int id) {
+        List<Writer> writers = writerRepository.getAll();
+        return writers.stream().noneMatch(writer -> writer.getId() == id);
+    }
 }
